Классы форматирования выделены в отдельный пакет.
