+++ engine/src/main/java/com/github/otymko/jos/runtime/format/NumberFormatter.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package com.github.otymko.jos.runtime;
+package com.github.otymko.jos.runtime.format;
 
 import lombok.Getter;
 import lombok.Setter;
