+++ engine/src/main/java/com/github/otymko/jos/runtime/machine/MachineInstance.java
@@ -15,7 +15,7 @@ import com.github.otymko.jos.module.ModuleImage;
 import com.github.otymko.jos.runtime.Arithmetic;
 import com.github.otymko.jos.runtime.IVariable;
 import com.github.otymko.jos.runtime.RuntimeContext;
-import com.github.otymko.jos.runtime.ValueFormatter;
+import com.github.otymko.jos.runtime.format.ValueFormatter;
 import com.github.otymko.jos.runtime.Variable;
 import com.github.otymko.jos.runtime.VariableReference;
 import com.github.otymko.jos.runtime.context.AttachableContext;
